# Changed AutoVectorize to use return by value for better performance